🚚 (Shorts-API): Cleaned up file structure
