+++ libs/api/shorts/api/feature/src/lib/api-shorts-api-feature.module.ts
@@ -26,7 +26,7 @@ import {
   CreateReportHandler,
   DeleteReportHandler,
 } from '@graduates/api/shorts/service/feature';
-import { ShortsResolver } from './api-shorts-resolver.resolver';
+import { ShortsResolver } from './shorts/api-shorts-resolver.resolver';
 import { ShortsService } from '@graduates/api/shorts/service/feature';
 import { PrismaService } from '@graduates/api/shared/services/prisma/data-access';
