+++ libs/api/shorts/service/feature/src/index.ts
@@ -1,7 +1,9 @@
-export * from './lib/api-shorts-service.service';
-export * from './lib/queries/api-shorts-query-handler.handler';
-export * from './lib/commands/api-shorts-command-handler.handler';
-export * from './lib/queries/api-short-tag-query-handler.handler';
-export * from './lib/commands/api-short-tag-command-handler.handler';
-export * from './lib/queries/api-short-report-query-handler.handler';
-export * from './lib/commands/api-short-report-command-handler.handler';
+export * from './lib/shorts/api-shorts-service.service';
+export * from './lib/short-tags/api-shorts-tags-service.service';
+export * from './lib/short-reports/api-shorts-reports-service.service';
+export * from './lib/shorts/queries/api-shorts-query-handler.handler';
+export * from './lib/shorts/commands/api-shorts-command-handler.handler';
+export * from './lib/short-tags/queries/api-short-tag-query-handler.handler';
+export * from './lib/short-tags/commands/api-short-tag-command-handler.handler';
+export * from './lib/short-reports/queries/api-short-report-query-handler.handler';
+export * from './lib/short-reports/commands/api-short-report-command-handler.handler';
