+++ libs/api/shorts/service/feature/src/lib/shorts/api-shorts-service.service.spec.ts
@@ -148,139 +148,4 @@ describe('ShortsService', () => {
       );
     });
   });
-
-  /**
-   * Test the findAllTags method
-   */
-  describe('findAllTags', () => {
-    const result = [tagMock];
-    it('should return an array of ShortTags', async () => {
-      jest
-        .spyOn(service, 'findAllTags')
-        .mockImplementation((): Promise<ShortTag[]> => Promise.resolve(result));
-
-      expect(await service.findAllTags()).toMatchObject(result);
-    });
-  });
-
-  /**
-   * Test the findTagsByShortId method
-   */
-  describe('findTagsByShortId', () => {
-    const result = [tagMock];
-    it('should return an array of ShortTags', async () => {
-      jest
-        .spyOn(service, 'findTagsByShortId')
-        .mockImplementation((): Promise<ShortTag[]> => Promise.resolve(result));
-
-      expect(await service.findTagsByShortId('1')).toMatchObject(result);
-    });
-  });
-
-  /**
-   * Test the createTag method
-   */
-  describe('createTag', () => {
-    it('should return a ShortTag', async () => {
-      jest
-        .spyOn(service, 'createTag')
-        .mockImplementation((): Promise<ShortTag> => Promise.resolve(tagMock));
-
-      expect(await service.createTag(tagCreateMock)).toMatchObject(tagMock);
-    });
-  });
-
-  /**
-   * Test the updateTags method
-   */
-  describe('updateTags', () => {
-    it('should return "success"', async () => {
-      jest.spyOn(service, 'updateTags').mockResolvedValue('success');
-
-      expect(await service.updateTags('tag', 'newTag')).toEqual('success');
-    });
-    it('should return "No tags updated"', async () => {
-      jest.spyOn(service, 'updateTags').mockResolvedValue('No tags updated');
-
-      expect(await service.updateTags('tag', 'newTag')).toEqual(
-        'No tags updated'
-      );
-    });
-  });
-
-  /**
-   * Test the updateTagByShortId method
-   */
-  describe('updateTagByShortId', () => {
-    it('should return "success"', async () => {
-      jest.spyOn(service, 'updateTagByShortId').mockResolvedValue('success');
-
-      expect(await service.updateTagByShortId('1', 'tag', 'newTag')).toEqual(
-        'success'
-      );
-    });
-    it('should return "Tag not found"', async () => {
-      jest
-        .spyOn(service, 'updateTagByShortId')
-        .mockResolvedValue('Tag not found');
-
-      expect(await service.updateTagByShortId('1', 'tag', 'newTag')).toEqual(
-        'Tag not found'
-      );
-    });
-  });
-
-  /**
-   * Test the deleteTags method
-   */
-  describe('deleteTags', () => {
-    it('should return "success"', async () => {
-      jest.spyOn(service, 'deleteTags').mockResolvedValue('success');
-
-      expect(await service.deleteTags('tag')).toEqual('success');
-    });
-    it('should return "No tags deleted"', async () => {
-      jest.spyOn(service, 'deleteTags').mockResolvedValue('No tags deleted');
-
-      expect(await service.deleteTags('tag')).toEqual('No tags deleted');
-    });
-  });
-
-  /**
-   * Test the deleteTagsByShortId method
-   */
-  describe('deleteTagsByShortId', () => {
-    it('should return "success"', async () => {
-      jest.spyOn(service, 'deleteTagsByShortId').mockResolvedValue('success');
-
-      expect(await service.deleteTagsByShortId('1')).toEqual('success');
-    });
-    it('should return "No tags deleted"', async () => {
-      jest
-        .spyOn(service, 'deleteTagsByShortId')
-        .mockResolvedValue('No tags deleted');
-
-      expect(await service.deleteTagsByShortId('1')).toEqual('No tags deleted');
-    });
-  });
-
-  /**
-   * Test the deleteTagByShortTag method
-   */
-  describe('deleteTagByShortTag', () => {
-    it('should return "success"', async () => {
-      jest.spyOn(service, 'deleteTagByShortTag').mockResolvedValue('success');
-
-      expect(await service.deleteTagByShortTag('1', 'tag')).toEqual('success');
-    });
-    it('should return "Tag not found"', async () => {
-      jest
-        .spyOn(service, 'deleteTagByShortTag')
-        .mockResolvedValue('Tag not found');
-
-      expect(await service.deleteTagByShortTag('1', 'tag')).toEqual(
-        'Tag not found'
-      );
-    });
-  });
 });
